pyproject.toml
[tool.poetry]
name = "streamlit-app"
version = "0.1.0"
description = ""
authors = ["Marlos Ribeiro <marlos.mgrb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.36.0"
pandas = "^2.2.2"
yfinance = "^0.2.41"
plotly = "^5.23.0"
pillow = "^10.4.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
